created Interfaces namespace and sorted scripts
