+++ Assets/Scripts/AttackLogic.cs
@@ -1,4 +1,5 @@
-﻿using Player;
+﻿using Interfaces;
+using Player;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
