+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Interfaces
+{
+    public interface IDamageable
+    {
+        void Hit(int damage);
+
+        void Die();
+    }
+}
+
