--- Assets/Scripts/Player/IDamageable.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IDamageable
-{
-    void Hit(int damage);
-
-    void Die();
-}
