+++ Assets/Scripts/Player/Stats.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using Interfaces;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
